Revert "nr-phy, nr-gnb-phy, nr-spectrum-phy: Resolving issue detected in NR-U simulations"
This reverts commit c34be62e3ff76519ad311f11155f62f5379e1e16.

Merge request that adds the reproducibility !37 testing has detected an issue with recently merged commit #c34be62e from !29 (merged) MR.
We need to revert back this #c34be62e commit until the issue is resolved. The issue is that after this commit, the uplink flows in cttc-3gpp-channel-nums-fdm are 0.

See issue 149.