nr: Remove misleading HarqEnabled flag from NrHelper and add test case
HarqEnabled flag is described as the flag that disables HARQ,
but what it does is to disable only the HARQ feedback. This is
misleading feature, not like the one implemented in the LTE module, where we can completely disable or enable HARQ logic.
In the NR module we have the flag ns3::NrMacSchedulerNs3::EnableHarqReTx that
enables/disables HARQ retransmissions. While full HARQ logic control like in LTE module is not implemented yet.

Co-authored-by: Gabriel Ferreira <gabrielcarvfer@gmail.com>